Added an ALTER for user_achievements.exponent
tracks internal commit 83b25eb545bf4f9d7417183a945303e7a96e121c

+++ plugins/Achievements/mysql_schema.sql
@@ -13,7 +13,7 @@ CREATE TABLE user_achievements (
         id mediumint(8) unsigned NOT NULL auto_increment,
         uid mediumint(8) unsigned NOT NULL default '0',
         aid mediumint(8) unsigned NOT NULL default '0',
-        exponent tinyint(1) unsigned NOT NULL default '0',
+        exponent smallint unsigned NOT NULL default '0',
         createtime datetime NOT NULL default '0000-00-00 00:00:00',
         PRIMARY KEY  (id),
         UNIQUE KEY achievement (uid,aid)
